+++ changelog.md
@@ -92,7 +92,7 @@
 - APP : iPad version back on the track ! 🎉
 - APP : Rename Fort-to-nite.
 - UI : You are now able to set favorites ! ❤️
-- 
+- UI : More tab with Fav, contact, licences row.